Progress Report
Right, so I've been tinkering with the object loading. I got it to work, though I couldn't manage to do it with any actual household items; I made it work with the same item we use in the learnopengl.com example, a nanosuit (which isn't terribly helpful at the moment, but baby steps). The main problem right now is that when we import an object into the scene, there's no telling how big it'll be or where it is with respect to the camera. It's either there or it's not, and it's a bitch trying to find the object. But anyway, I managed to fix it such that we have an object successfully loaded into the scene. I made some minor changes to the script you wrote, mostly aesthetic and pragmatic stuff, you'll see. I've sent you the other objects I was trying to work with, too. I got these models from TurboSquid.com, it seems really perfect for this, but I don't quite have a handle on how to transfer the files. The .mtl files are meant to contain information in such a way as to inform the system of where the textures and such are, but I think there's something about the process that I'm doing wrong because even when I could see the couches, the textures were absent.