Progress Report
Made changes to ModelClass and MeshClass - please update your versions with these ones.
Also made a great many changes to the object models, but you shouldn't need to change any filepaths.
I've implemented the user select functionality, but I need your help to look it over and possibly modify a thing or two; it works, but I'm not sure if I'm quite doing it the right way. My confusion is with the whole local coordinates vs world coordinates; we can talk about it a bit later, tonight. You should have everything you need to work on the collision; see ModelObject.hpp's BoundingBox struct and how it's used in the ModelObject class for more info.