Lights Fully implemented now
I made lots of progress :D
- lighting now includes specular component for all ground, wall and
objects
- Locations of lights were fixed
- Added a cube to represent the lightbulb  (to make location of light
easier)
- When pressing 1,2,3 will turn off kitchen, livingroom and bedroom
lights respectively
-wall collision now detectable
(cannot go through any of the walls, but ceiling and floor not
implemented)

NOTE: I just looked through the project requirements and it seems that
all thats left to implement is collision and control of furniture.
I just reread some of the requirements... There is two things that might
need to be changed BUT its a quick fix and if we aggree that the wording
implies what I think it does.

1. In the second requirement, it says that "Each room ceiling must have
at least one light, and some of the walls must have windows"
SO if I understand correctly we need ceiling lights for the living room
and the bedroom (unless we decide to argue and say that we do have
lights)
AND the word "windows" is bugging me so to be safe I could easily add
more windows just be sure

2. This one really bugs me :(
"Use different textures for walls, floors, and ceilings."
Do they mean that for each room, the walls, floors and ceilings must
have different textures or different textures for the walls, floors and
ceilings?
Its really bugging me... Well discuss this tonight.