Made a change in the transformModelObject function
Now I noticed that the objects were being moved based on x, y, z axis
and while its fine for the y axis, figuring out which way the x and z
axis is when the directions constantly change based on the camera
orientation is annoying (its doable but inefficient)
SO I made a few changes that will allow to use up and down key to move
the object forward and backward, and left-right to move the object to
the left and to the right.
Its much easier this way to figure out which way the object is going to.