Double-checking
Sup cutie?

Yo, would you please go over this .zip and make sure I've not left anything out with respect to includes or whatever? I went over it a solid five times, but I'd appreciate a second pair of eyes on this. The submission time is now confirmed as being Tuesday the 20th, so today, at 2355. We still have the time, may as well ensure all's where it should be.
Ignore all but the .zip, its the only relevant document to this push.